Part 1
We need to calculate the constraining dimension of the browser window for the image. Will the height or the width decide what size to draw the image?
Once we know the constraining dimension, we need to calculate an offset for the other larger dimension so the image is centered in the browser window.